requirements.txt
# 基础依赖
tensorflow>=2.0.0
pillow>=8.0.0
python-magic>=0.4.0
matplotlib>=3.0.0
numpy>=1.19.0
pandas>=1.1.0
psutil>=5.8.0
xxhash>=2.0.0
schedule>=1.1.0

# GUI依赖
tkinter  # 通常随Python安装
PyQt5>=5.15.0  # 可选的GUI后端

# 开发依赖
pytest>=6.0.0
pytest-cov>=2.10.0
black>=20.8b1
flake8>=3.8.0
mypy>=0.800
jupyter>=1.0.0

# 生产依赖
gunicorn>=20.0.4
sentry-sdk>=0.19.0